pyproject.toml
[project]
name = "ml-3"
version = "0.1.0"
description = ""
authors = [
    {name = "thanakara",email = "dngus1717@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10,<3.12"

dependencies = [
    "matplotlib (>=3.8.1,<3.9.0)",
    "numpy (>=1.26.2,<1.27.0)",
    "pandas (>=2.1.3,<2.2.0)",
]

[tool.poetry.group.fw.dependencies]
hydra-core = ">=1.3,<2.0"
scikit-learn = ">=1.3.2,<1.4.0"
tensorflow-io-gcs-filesystem = "0.31.0"
# tensorflow-intel = "2.14.0"
tensorflow = "2.14.0"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[project.scripts]
run = "ml_3.__main__:main"

